Animation fixes. Refactoring.

+++ Match-3/Assets/Scripts/Components/Game/Events/UpdateTurnTimerViewRequest.cs
@@ -2,6 +2,6 @@
 {
     public struct UpdateTurnTimerViewRequest
     {
-        public float TimeRamain;
+        public int TimeRamain;
     }
 }
